Separate Hero Feature into Functional Components
